+++ js/tracker.php
@@ -48,7 +48,13 @@ if (!defined("PIWIK_KEEP_JS_TRACKER_COMMENT")
     $byteStart = 371; // length of comment header in bytes
 }
 
-$environment = new \Piwik\Application\Environment(null);
+class Validator {
+    public function validate() {}
+}
+$validator = new Validator();
+$environment = new \Piwik\Application\Environment(null, array(
+    'Piwik\Application\Kernel\EnvironmentValidator' => $validator
+));
 $environment->init();
 
 ProxyHttp::serverStaticFile($file, "application/javascript; charset=UTF-8", $daysExpireFarFuture, $byteStart, $byteEnd);