+++ plugins/CoreHome/CoreHome.php
@@ -109,6 +109,7 @@ class CoreHome extends \Piwik\Plugin
         $stylesheets[] = "plugins/CoreHome/angularjs/quick-access/quick-access.directive.less";
         $stylesheets[] = "plugins/CoreHome/angularjs/selector/selector.directive.less";
         $stylesheets[] = "plugins/CoreHome/angularjs/reporting-page/reportingpage.directive.less";
+        $stylesheets[] = "plugins/CoreHome/angularjs/report-export/reportexport.popover.less";
         $stylesheets[] = "plugins/CoreHome/angularjs/widget-bydimension-container/widget-bydimension-container.directive.less";
         $stylesheets[] = "plugins/CoreHome/angularjs/progressbar/progressbar.directive.less";
         $stylesheets[] = "plugins/CoreHome/angularjs/date-range-picker/date-range-picker.component.less";
@@ -191,6 +192,7 @@ class CoreHome extends \Piwik\Plugin
         $jsFiles[] = "plugins/CoreHome/angularjs/common/directives/dropdown-button.js";
         $jsFiles[] = "plugins/CoreHome/angularjs/common/directives/select-on-focus.js";
         $jsFiles[] = "plugins/CoreHome/angularjs/common/directives/side-nav.js";
+        $jsFiles[] = "plugins/CoreHome/angularjs/common/directives/string-to-number.js";
 
         $jsFiles[] = "plugins/CoreHome/angularjs/piwikApp.js";
         $jsFiles[] = "plugins/CoreHome/angularjs/anchorLinkFix.js";
@@ -228,6 +230,8 @@ class CoreHome extends \Piwik\Plugin
         $jsFiles[] = "plugins/CoreHome/angularjs/widget/widget.directive.js";
 
         $jsFiles[] = "plugins/CoreHome/angularjs/popover-handler/popover-handler.directive.js";
+
+        $jsFiles[] = "plugins/CoreHome/angularjs/report-export/reportexport.directive.js";
 
         $jsFiles[] = "plugins/CoreHome/angularjs/reporting-page/reportingpage.controller.js";
         $jsFiles[] = "plugins/CoreHome/angularjs/reporting-page/reportingpage-model.js";
@@ -279,6 +283,7 @@ class CoreHome extends \Piwik\Plugin
 
     public function getClientSideTranslationKeys(&$translationKeys)
     {
+        $translationKeys[] = 'General_Export';
         $translationKeys[] = 'General_InvalidDateRange';
         $translationKeys[] = 'General_Loading';
         $translationKeys[] = 'General_Show';
@@ -383,6 +388,7 @@ class CoreHome extends \Piwik\Plugin
         $translationKeys[] = 'General_YourChangesHaveBeenSaved';
         $translationKeys[] = 'General_LearnMore';
         $translationKeys[] = 'General_ChooseDate';
+        $translationKeys[] = 'General_ReadThisToLearnMore';
         $translationKeys[] = 'CoreHome_UndoPivotBySubtable';
         $translationKeys[] = 'CoreHome_PivotBySubtable';
         $translationKeys[] = 'General_LearnMore';
@@ -400,11 +406,28 @@ class CoreHome extends \Piwik\Plugin
         $translationKeys[] = 'General_DoubleClickToChangePeriod';
         $translationKeys[] = 'General_Apply';
         $translationKeys[] = 'General_Period';
+        $translationKeys[] = 'CoreHome_EnterZenMode';
+        $translationKeys[] = 'CoreHome_ExitZenMode';
         $translationKeys[] = 'CoreHome_ShortcutZenMode';
         $translationKeys[] = 'CoreHome_ShortcutSegmentSelector';
         $translationKeys[] = 'CoreHome_ShortcutWebsiteSelector';
         $translationKeys[] = 'CoreHome_ShortcutCalendar';
         $translationKeys[] = 'CoreHome_ShortcutSearch';
         $translationKeys[] = 'CoreHome_ShortcutHelp';
+        $translationKeys[] = 'CoreHome_StandardReport';
+        $translationKeys[] = 'CoreHome_ReportWithMetadata';
+        $translationKeys[] = 'CoreHome_ReportType';
+        $translationKeys[] = 'CoreHome_RowLimit';
+        $translationKeys[] = 'CoreHome_ExportFormat';
+        $translationKeys[] = 'CoreHome_FlattenReport';
+        $translationKeys[] = 'CoreHome_CustomLimit';
+        $translationKeys[] = 'CoreHome_ExpandSubtables';
+        $translationKeys[] = 'CoreHome_HomeShortcut';
+        $translationKeys[] = 'CoreHome_PageUpShortcutDescription';
+        $translationKeys[] = 'CoreHome_EndShortcut';
+        $translationKeys[] = 'CoreHome_PageDownShortcutDescription';
+        $translationKeys[] = 'CoreHome_MacPageUp';
+        $translationKeys[] = 'CoreHome_MacPageDown';
+
     }
 }